backend/pyproject.toml
[project]
name = "agentx-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "anthropic>=0.43.0",
    "langfuse>=2.50.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "python-multipart>=0.0.18",
    "resend>=2.0.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-magic>=0.4.27",
    "structlog>=24.4.0",
    "Pillow>=11.0.0",
    "email-validator>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0.0", "pytest-asyncio>=0.24.0"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
